Addressing some changes brittney recommended and preparing for new view test
